Retry in case of EPIPE.
This patch, in conjunction with a non-zero value for Retries (which
defaults to 5), will allow Gossie connection pools to tolerate Cassandra
nodes restarted underneath the pool with no application impact.

Thrift masks EPIPE pretty aggressively but as the tests indicate, both
reads and writes produce errors that are distinct enough to match to
detect this condition reliably.